Move MainActivity to .ui Dagger init
