+++ app/src/main/java/ru/ft/giphyapp/di/AppComponent.kt
@@ -1,0 +1,9 @@
+package ru.ft.giphyapp.di
+
+import dagger.Component
+import ru.ft.giphyapp.ui.home.HomeFragment
+
+@Component(modules = [AppModule::class])
+interface AppComponent {
+    fun inject(homeFragment: HomeFragment)
+}
