+++ app/src/main/java/ru/ft/giphyapp/di/AppModule.kt
@@ -1,0 +1,10 @@
+package ru.ft.giphyapp.di
+
+import dagger.Module
+
+@Module
+interface AppModule {
+
+//    @Binds
+//    fun bindRepository(repositoryImpl: RepositoryImpl): Repository
+}
