+++ app/src/main/java/ru/ft/giphyapp/ui/MainActivity.kt
@@ -1,5 +1,6 @@
-package ru.ft.giphyapp
+package ru.ft.giphyapp.ui
 
 import androidx.appcompat.app.AppCompatActivity
+import ru.ft.giphyapp.R
 
 class MainActivity : AppCompatActivity(R.layout.activity_main)
