+++ app/src/main/java/ru/ft/giphyapp/ui/home/HomeViewModel.kt
@@ -1,6 +1,8 @@
 package ru.ft.giphyapp.ui.home
 
 import androidx.lifecycle.ViewModel
+import javax.inject.Inject
 
-class HomeViewModel : ViewModel() {
+class HomeViewModel @Inject constructor(
+) : ViewModel() {
 }
