+++ app/src/main/res/layout/activity_main.xml
@@ -8,4 +8,4 @@
     android:layout_height="match_parent"
     app:defaultNavHost="true"
     app:navGraph="@navigation/nav_graph"
-    tools:context=".MainActivity" />
+    tools:context=".ui.MainActivity" />
